+++ test/apps/__init__.py
@@ -28,20 +28,20 @@ AvailableEndpoints = CSVOption(Endpoint)
 
 @click.command()
 @click.argument("port", type=int)
-@click.option("--endpoints", type=AvailableEndpoints)
+@click.option("--operations", type=AvailableEndpoints)
 @click.option("--spec", type=click.Choice(["openapi2", "openapi3", "graphql"]), default="openapi2")
 @click.option("--framework", type=click.Choice(["aiohttp", "flask"]), default="aiohttp")
-def run_app(port: int, endpoints: List[Endpoint], spec: str, framework: str) -> None:
+def run_app(port: int, operations: List[Endpoint], spec: str, framework: str) -> None:
     if spec == "graphql":
         app = _graphql.create_app()
         app.run(port=port)
     else:
-        if endpoints is not None:
-            prepared_endpoints = tuple(endpoint.name for endpoint in endpoints)
-            if "all" in prepared_endpoints:
-                prepared_endpoints = tuple(endpoint.name for endpoint in Endpoint if endpoint.name != "all")
+        if operations is not None:
+            prepared_operations = tuple(endpoint.name for endpoint in operations)
+            if "all" in prepared_operations:
+                prepared_operations = tuple(endpoint.name for endpoint in Endpoint if endpoint.name != "all")
         else:
-            prepared_endpoints = tuple(
+            prepared_operations = tuple(
                 endpoint.name
                 for endpoint in Endpoint
                 if endpoint.name not in INVALID_ENDPOINTS and endpoint.name != "all"
@@ -53,10 +53,10 @@ def run_app(port: int, endpoints: List[Endpoint], spec: str, framework: str) -> 
             bold=True,
         )
         if framework == "aiohttp":
-            app = _aiohttp.create_openapi_app(prepared_endpoints, version)
+            app = _aiohttp.create_openapi_app(prepared_operations, version)
             web.run_app(app, port=port)
         elif framework == "flask":
-            app = _flask.create_openapi_app(prepared_endpoints, version)
+            app = _flask.create_openapi_app(prepared_operations, version)
             app.run(port=port)